requirements.txt
# Flask and extensions
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.7.1
flask-restx

# Database
SQLAlchemy==2.0.32
Alembic==1.13.2
psycopg2-binary>=2.9,<3.0
Mako==1.0.7

# Core WSGI/Web
Werkzeug
Jinja2
itsdangerous
click

# Serialization and Data
jsonschema>=4.0.0,<5.0.0
aniso8601==3.0.0
python-dateutil==2.7.0
pytz

# Security/Crypto
bcrypt==3.1.4
cffi==1.15.0
pycparser==2.18
PyJWT>=2.4.0

# Utilities
python-dotenv
python-editor==1.0.3
requests==2.28.1
six==1.11.0
MarkupSafe
boto3==1.24.25